Modified PriceHistory and DBPriceHistory to use long for ItemID.
Implemented DBPriceHistoryQuery to use PriceHistoryQuery interface. This
class provides methods to add and get price history items from the DB.

DBPriceHistoryQueryTest uses JUnit to test the DBPriceHistoryQuery class.

+++ src/main/java/edu/cpp/cs580/Database/Objects/Interfaces/PriceHistory.java
@@ -1,14 +1,13 @@
 package edu.cpp.cs580.Database.Objects.Interfaces;
 
-import java.math.BigInteger;
-import java.util.Date;
+import java.sql.Date;
 
 public interface PriceHistory {
-	public BigInteger getItemID();
+	public long getItemID();
 	public double getPrice();
 	public Date getPriceDate();
 	public int getStoreID();
-	public void setItemID(BigInteger id);
+	public void setItemID(long id);
 	public void setPrice(double pr);
 	public void setPriceDate(Date date);
 	public void setStoreID(int id);
